requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
scikit-learn>=1.5.0
pandas==2.3.3
numpy==2.1.0
joblib==1.3.2
Werkzeug==3.0.1
PyPDF2==3.0.1
pymongo==4.6.1
python-dotenv==1.0.0